backend/requirements.txt
fastapi
uvicorn
requests
pydantic
google-generativeai
scaledown